add unit tests (internal and external)
