+++ src/_common/Quotes/Quote.Converters.cs
@@ -9,7 +9,7 @@ public static partial class QuoteUtility
 {
     private static readonly CultureInfo NativeCulture = Thread.CurrentThread.CurrentUICulture;
 
-    /* LISTS */
+    /* STANDARD DECIMAL QUOTES */
 
     // convert TQuotes to basic double tuple list
     /// <include file='./info.xml' path='info/type[@name="UseCandlePart"]/*' />
@@ -26,7 +26,7 @@ public static partial class QuoteUtility
         where TQuote : IQuote => quotes
             .ToSortedList()
             .ToCollection();
-        
+
     internal static List<TQuote> ToSortedList<TQuote>(
         this IEnumerable<TQuote> quotes)
         where TQuote : IQuote => quotes
