+++ src/_common/Quotes/Quote.Exceptions.cs
