+++ src/_common/Quotes/Quote.Validation.cs
@@ -1,5 +1,3 @@
-using System.Globalization;
-
 namespace Skender.Stock.Indicators;
 
 // HISTORICAL QUOTES VALIDATION
