+++ tests/external/Test.CustomIndicator.cs
@@ -183,6 +183,29 @@ public class CustomIndicatorTests
     }
 
     [TestMethod]
+    public void QuoteToSortedList()
+    {
+        IEnumerable<Quote> quotes = TestData.GetMismatch();
+
+        Collection<Quote> h = quotes.ToSortedCollection();
+
+        // proper quantities
+        Assert.AreEqual(502, h.Count);
+
+        // check first date
+        DateTime firstDate = DateTime.ParseExact("01/18/2016", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(firstDate, h[0].Date);
+
+        // check last date
+        DateTime lastDate = DateTime.ParseExact("12/31/2018", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(lastDate, h.LastOrDefault().Date);
+
+        // spot check an out of sequence date
+        DateTime spotDate = DateTime.ParseExact("03/16/2017", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(spotDate, h[50].Date);
+    }
+
+    [TestMethod]
     public void TupleNaN()
     {
         List<MyResult> r = tupleNanny
