+++ tests/indicators/_common/Quotes/Test.Quote.Converters.cs
@@ -8,6 +8,52 @@ namespace Internal.Tests;
 public class QuoteUtility : TestBase
 {
     [TestMethod]
+    public void QuoteToSortedCollection()
+    {
+        IEnumerable<Quote> quotes = TestData.GetMismatch();
+
+        Collection<Quote> h = quotes.ToSortedCollection();
+
+        // proper quantities
+        Assert.AreEqual(502, h.Count);
+
+        // check first date
+        DateTime firstDate = DateTime.ParseExact("01/18/2016", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(firstDate, h[0].Date);
+
+        // check last date
+        DateTime lastDate = DateTime.ParseExact("12/31/2018", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(lastDate, h.LastOrDefault().Date);
+
+        // spot check an out of sequence date
+        DateTime spotDate = DateTime.ParseExact("03/16/2017", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(spotDate, h[50].Date);
+    }
+
+    [TestMethod]
+    public void QuoteToSortedList()
+    {
+        IEnumerable<Quote> quotes = TestData.GetMismatch();
+
+        List<Quote> h = quotes.ToSortedList();
+
+        // proper quantities
+        Assert.AreEqual(502, h.Count);
+
+        // check first date
+        DateTime firstDate = DateTime.ParseExact("01/18/2016", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(firstDate, h[0].Date);
+
+        // check last date
+        DateTime lastDate = DateTime.ParseExact("12/31/2018", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(lastDate, h.LastOrDefault().Date);
+
+        // spot check an out of sequence date
+        DateTime spotDate = DateTime.ParseExact("03/16/2017", "MM/dd/yyyy", EnglishCulture);
+        Assert.AreEqual(spotDate, h[50].Date);
+    }
+
+    [TestMethod]
     public void QuoteToTuple()
     {
         DateTime d = DateTime.Parse("5/5/2055", EnglishCulture);
